Change Parrot_assert condition argument from int to long, in an attempt to remove warnings on platforms with 64-bit pointers but 32-bit ints.
git-svn-id: https://svn.parrot.org/parrot/trunk@25741 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ include/parrot/exceptions.h
@@ -193,7 +193,7 @@ void new_internal_exception(PARROT_INTERP)
 PARROT_API
 PARROT_DOES_NOT_RETURN_WHEN_FALSE
 void Parrot_assert(
-    int condition,
+    long condition,
     ARGIN(const char *condition_string),
     ARGIN(const char *file),
     unsigned int line)
@@ -304,7 +304,7 @@ void rethrow_c_exception(PARROT_INTERP)
 #ifdef NDEBUG
 #  define PARROT_ASSERT(x) ((void)0)
 #else
-#  define PARROT_ASSERT(x) Parrot_assert((int)(x), #x, __FILE__, __LINE__)
+#  define PARROT_ASSERT(x) Parrot_assert((long)(x), #x, __FILE__, __LINE__)
 #endif
